+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robot.java
@@ -18,19 +18,13 @@ import org.firstinspires.ftc.teamcode.mechanisms.MecanumDrive;
 public class Robot {
     public Globals globals = Globals.getInstance();
     public MecanumDrive mecanumDrive = new MecanumDrive();
-    public ManipulatorArm manipulatorArm = new ManipulatorArm();
-//    public StageTwoAscentArms stageTwoAscentArms = new StageTwoAscentArms();
-    public StageOneAscentArms stageOneAscentArms = new StageOneAscentArms();
     public LimelightCamera camera = new LimelightCamera();
     public OpticalOdometry opticalOdometry = new OpticalOdometry();
 
     public void init(HardwareMap hardwareMap){
         mecanumDrive.init(hardwareMap);
-        manipulatorArm.init(hardwareMap);
-        //stageTwoAscentArms.init(hardwareMap);
-        stageOneAscentArms.init(hardwareMap);
-        camera.init(hardwareMap);
-        opticalOdometry.init(hardwareMap);
+        //camera.init(hardwareMap);
+        //opticalOdometry.init(hardwareMap);
     }
 
     public Position getPosition(){
